Localize time.
